requirements.txt
flask
flask-cors
python-dotenv
pandas
requests
sqlalchemy
pymysql
cryptography
